Add stats (M300) and runtime modifiable configuration options.

+++ demitasse/gcode_parse.h
@@ -2,10 +2,10 @@
 #define	_GCODE_PARSE_H
 
 #ifndef _WIN32
-#include	<stdint.h>
+#include <stdint.h>
 #endif
 
-#include	"motion_planner.h"
+#include "motion_planner.h"
 
 // wether to insist on N line numbers
 // if not defined, N's are completely ignored
@@ -137,7 +137,7 @@ typedef struct {
 	uint16_t					M;				///< M command number
 	TARGET						target;		///< target position: X, Y, Z, E and F
 
-	int16_t						S;				///< S word (various uses)
+	int32_t						S;				///< S word (various uses)
 	uint16_t					P;				///< P word (various uses)
 
 	uint8_t						T;				///< T word (tool index)
